gcp/gcf/datastore_test/requirements.txt
google-cloud-bigquery == 2.7.0
google-cloud-datastore == 2.1.0
flask